requirements.txt
django==1.10
djangorestframework==3.4.1
gunicorn==19.6.0
psycopg2==2.6.2
requests==2.11.1
raven
newrelic

django-cors-middleware
django-filter
django-crispy-forms
django_markdown2
faker
